Initial commit for Advent of Code 2025
The setup script may be used to create a worktree branch for a day's puzzle:

```
./setup-puzzle.sh 2 `cat .cookie`
```

Where "2" is the day number and ".cookie" is the file where the
session cookie is stored for puzzle input.

https://adventofcode.com/2025

To the King of the ages, immortal, invisible, the only God, be honor and glory forever and ever. Amen.