Updates solution to part 2
This update takes a more memory efficient approach by simply examining
the boundaries of ranges for overlap instead of each elements of ranges
for overlap. Because the ranges are inclusive, it is not necessary to
examine each individual element.

> That's the right answer! You are one gold star closer to decorating the North Pole.
You have completed Day 5!

Your puzzle answer was `344771884978261`.

Refer to previous commit fc0375b

To the King of the ages, immortal, invisible, the only God, be honor and glory forever and ever. Amen.