Adds Cargo Rust project with part 1 solution
> Your puzzle answer was `1626`.
That's the right answer! You are one gold star closer to decorating the North Pole.

I am not satisfied with the implementation. However, because it worked,
I'll see how I can optimize it for part 2.

Refer to previous commit 8f8b22b

To the King of the ages, immortal, invisible, the only God, be honor and glory forever and ever. Amen.